+++ src/OaiPmh/HarvesterCommand.php
@@ -35,6 +35,7 @@ use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 use VuFindHarvest\ConsoleOutput\ConsoleWriter;
 use VuFindHarvest\ConsoleOutput\WriterAwareTrait;
+use VuFindHarvest\Exception\OaiException;
 
 /**
  * OAI-PMH Harvest Tool (Symfony Console Command)
@@ -332,7 +333,7 @@ class HarvesterCommand extends Command
         }
 
         // Loop through all the settings and perform harvests:
-        $processed = $skipped = 0;
+        $processed = $skipped = $errors = 0;
         foreach ($allSettings as $target => $baseSettings) {
             $settings = $this->updateSettingsWithConsoleOptions(
                 $input, $baseSettings
@@ -345,8 +346,14 @@ class HarvesterCommand extends Command
             try {
                 $this->harvestSingleRepository($input, $output, $target, $settings);
             } catch (\Exception $e) {
-                $this->writeLine($e->getMessage());
-                return 1;
+                if ($e instanceof OaiException
+                    && strtolower($e->getOaiCode()) == 'norecordsmatch'
+                ) {
+                    $this->writeLine("No new records found.");
+                } else {
+                    $this->writeLine($e->getMessage());
+                    $errors++;
+                }
             }
             $processed++;
         }
@@ -356,6 +363,13 @@ class HarvesterCommand extends Command
             $this->writeLine(
                 'No valid settings found; '
                 . 'please set url and metadataPrefix at minimum.'
+            );
+            return 1;
+        }
+        if ($errors > 0) {
+            $this->writeLine(
+                "Completed with {$errors} error(s) -- "
+                . "{$processed} source(s) processed."
             );
             return 1;
         }