+++ tests/unit-tests/src/VuFindTest/Exception/OaiExceptionTest.php
@@ -1,0 +1,60 @@
+<?php
+
+/**
+ * OAI exception test
+ *
+ * PHP version 7
+ *
+ * Copyright (C) Villanova University 2021.
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License version 2,
+ * as published by the Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
+ *
+ * @category VuFind
+ * @package  Tests
+ * @author   Demian Katz <demian.katz@villanova.edu>
+ * @license  http://opensource.org/licenses/gpl-2.0.php GNU General Public License
+ * @link     https://vufind.org/wiki/development
+ */
+namespace VuFindTest\Harvest\ConsoleOutput;
+
+use VuFindHarvest\Exception\OaiException;
+
+/**
+ * OAI exception test
+ *
+ * @category VuFind
+ * @package  Tests
+ * @author   Demian Katz <demian.katz@villanova.edu>
+ * @license  http://opensource.org/licenses/gpl-2.0.php GNU General Public License
+ * @link     https://vufind.org/wiki/development
+ */
+class OaiExceptionTest extends \PHPUnit\Framework\TestCase
+{
+    /**
+     * Test exception
+     *
+     * @return void
+     */
+    public function testException()
+    {
+        $exception = new OaiException('code', 'message');
+        $this->assertEquals('code', $exception->getOaiCode());
+        $this->assertEquals('message', $exception->getOaiMessage());
+        $this->assertEquals(
+            'OAI-PMH error -- code: code, value: message', $exception->getMessage()
+        );
+        $this->assertEquals(0, $exception->getCode());
+        $this->assertNull($exception->getPrevious());
+    }
+}